requirements.txt
gitpython==3.1.30
h5py==3.8.0
matplotlib==3.6.3
minisom==2.3.0
numpy==1.24.1
pandas==1.5.3
Pillow==9.4.0
pingouin==0.5.3
pyyaml==6.0
scikit-image==0.19.3
scikit-learn==1.2.1
scipy==1.10.0
shapely==2.0.0
torch-lucent==0.1.8
tqdm==4.64.1
xarray==0.17.0